+++ Dockerfile
@@ -7,8 +7,6 @@ RUN docker-php-ext-install pdo_pgsql
 COPY ./configs/php/php.ini /usr/local/etc/php/
 COPY --from=composer:latest /usr/bin/composer /usr/bin/composer
 
-COPY . .
-
 # 参考
 # https://qiita.com/YAJIMA/items/68de1bdeb71a921a718d
 RUN a2enmod rewrite \
